+++ spec/loan_creator/in_fine_spec.rb
@@ -37,6 +37,20 @@ describe LoanCreator::InFine do
       include_examples('valid lender timetable', loan_type, scenario, initial_values, true)
     end
 
+    context 'given a scenario with term dates' do
+      term_dates = ['2021-12-10','2021-12-15', '2022-01-15', '2022-02-15',
+                  '2022-03-15', '2022-04-15', '2022-05-15', '2022-06-15',
+                  '2022-07-15', '2022-08-15', '2022-09-15', '2022-10-15',
+                  '2022-11-15', '2022-12-15', '2023-01-15', '2023-02-15',
+                  '2023-03-15', '2023-04-15', '2023-05-15', '2023-06-15',
+                  '2023-07-15', '2023-08-15', '2023-09-15', '2023-10-15'
+                ]
+
+      scenario = ['month', '100000', '12', '2021-11-15', '24', '0']
+
+      include_examples('valid lender timetable', loan_type, scenario, {}, false, term_dates)
+    end
+
     context 'given a scenario with due_interests' do
       initial_values = {
         paid_capital: 0,