+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.9.0
+-------------------------
+- add possibility to pass the argument `term_dates` for `LoanCreator::Bullet`, `LoanCreator::InFine` and `LoanCreator::Linear`
+- `term_dates` option is based on `realistic_durations` and allow to compute terms custom date to custom date
+
 v0.8.2
 -------------------------